+++ M01/Assignment/ReadingList/app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#008577</color>
-    <color name="colorPrimaryDark">#00574B</color>
-    <color name="colorAccent">#D81B60</color>
+    <color name="colorPrimary">#455A64</color>
+    <color name="colorPrimaryDark">#1C313A</color>
+    <color name="colorAccent">#718792</color>
 </resources>
